requirements.txt
boto3==1.38.43
botocore==1.38.43
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
git-filter-repo==2.47.0
jmespath==1.0.1
joblib==1.5.1
kafka-python==2.2.13
kiwisolver==1.4.8
matplotlib==3.10.3
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pillow==11.2.1
py4j==0.10.9.9
pyparsing==3.2.3
pyspark==4.0.0
python-dateutil==2.9.0.post0
regex==2024.11.6
s3transfer==0.13.0
six==1.17.0
textblob==0.19.0
tqdm==4.67.1
urllib3==2.5.0
